% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ageify}
\alias{ageify}
\title{Calculate Age (or Time Difference) from Dates}
\usage{
ageify(DoEvent, DoBirth, format = "years")
}
\arguments{
\item{DoEvent}{a vector of class 'Date'}

\item{DoBirth}{a vector of class 'Date'}

\item{format}{unit in which the output should be returned. Options are \code{"years"} (default), 
\code{"months"} and \code{"days"}.}
}
\value{
Returns a numerical vector of the time difference between the two input vector Dates,
 calculated as \code{DoEvent - DoBirth}. Output can be given in units of years (default), 
 months or days, specified in \code{format}.
}
\description{
Calculates the difference between two dates and returns a numeric value (Age).
}
\examples{
#Calculate age of onset (from ResearchR package).
Age_onset <- ageify(d4$DoO, d4$DoB)

# For data.table output in ResearchR package
library(data.table)
d4[, c("AoO", "AoDiag") := lapply(.SD, ageify, DoBirth=DoB), 
   .SDcols=c("DoO","DoDiag")]


}
