\name{center.var}
\alias{center.var}
\title{
Center variables by mean for grouping variable
}
\description{}
\usage{
center.var(df, to.scale, scale.by)
}
\arguments{
  \item{df}{}
  \item{to.scale}{}
  \item{scale.by}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
