\name{cluster.lmax.grow}
\alias{cluster.lmax.grow}
\title{
Cluster based on local minima/maxima then grow to local maxima/minima
}
\description{}
\usage{
cluster.lmax.grow(tmap, tvol,pmap, pvol,mask, mvol,alpha = 0.05,lmax.radius = NULL,min.size = 10,min.size.action = "merge",save.dir,file.name = NULL,verbose = FALSE)
}
\arguments{
  \item{tmap}{}
  \item{tvol}{}
  \item{pmap}{}
  \item{pvol}{}
  \item{mask}{}
  \item{mvol}{}
  \item{alpha}{}
  \item{lmax.radius}{}
  \item{min.size}{}
  \item{min.size.action}{}
  \item{save.dir}{}
  \item{file.name}{}
  \item{verbose}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
