\name{cluster.lmax}
\alias{cluster.lmax}
\title{
Cluster based on effect sign, p value, and proximity to local minima/maxima
}
\description{}
\usage{
cluster.lmax(tmap, tvol, pmap, pvol, mask, mvol, alpha = 0.05, radius = 5, save.dir, file.name = NULL)
}
\arguments{
  \item{tmap}{}
  \item{tvol}{}
  \item{pmap}{}
  \item{pvol}{}
  \item{mask}{}
  \item{mvol}{}
  \item{alpha}{}
  \item{radius}{}
  \item{save.dir}{}
  \item{file.name}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
