\name{cluster.n.vol.1}
\alias{cluster.n.vol.1}
\title{Convert Multivolume to single}
\description{}
\usage{
  cluster.n.vol.1(nii.file, save.dir=NULL, file.name=NULL)
}
\arguments{
  \item{nii.file}{}
  \item{save.dir}{}
  \item{file.name}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}