\name{cluster.lmax}
\alias{cluster.lmax}
\title{
Cluster based on effect sign, p value, and proximity to local minima/maxima
}
\description{}
\usage{
cluster.valley(tmap, tvol, pmap, pvol, mask, mvol, alpha = 0.05, min.size = 10, save.dir, file.prefix)
}
\arguments{
  \item{tmap}{}
  \item{tvol}{}
  \item{pmap}{}
  \item{pvol}{}
  \item{mask}{}
  \item{mvol}{}
  \item{alpha}{}
  \item{min.size}{}
  \item{save.dir}{}
  \item{file.prefix}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
