\name{convolve.paradigm}
\alias{convolve.paradigm}
\title{Convolve Paradigm With Canonical HRF (from AFNI)}
\description{}
\usage{
  convolve.paradigm(data.onsets, scan.params, conv.type, save.dir, hrf.precision, ...)
}
\arguments{
  \item{data.onset}{}
  \item{scan.params}{}
  \item{conv.type}{}
  \item{save.dir}{}
  \item{hrf.precision}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
