\name{deconvolve}
\alias{deconvolve}
\title{temporary title}
\description{}
\usage{
deconvolve(data.4d, data.onsets, scan.params, data.mask, save.dir=getwd(), num.cores, incl.trend = c(TRUE, FALSE)) 
}
\arguments{
  \item{data.4d}{}
  \item{data.onsets}{}
  \item{scan.params}{}
  \item{data.mask}{}
  \item{save.dir}{}
  \item{num.cores}{}
  \item{incl.trend}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}