% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{DEFINITION_ECAS_BHV}
\alias{DEFINITION_ECAS_BHV}
\title{Clean up ECAS behaviour fields}
\usage{
DEFINITION_ECAS_BHV(data = long3$ECAS)
}
\arguments{
\item{data}{een data.table met longitudinale ECAS data. Default \code{long3$ECAS}}
}
\value{
Dezelfde data.table met longitudinale ECAS data, waarbij foutieve velden op NA zijn 
gezet en rijen zonder ECAS data zijn verwijderd.
}
\description{
Door het vinkjessysteem in Progeny zullen veel velden voor ECAS behaviour subscores standaard 
op \code{N} staan. Deze functie zet alle velden op \code{NA} indien:
1) Alle behaviour sumscores \code{NA} zijn EN 
2) Alle subscores \code{NA} of \code{N} zijn.
}
