\name{draw.summary}
\alias{draw.summary}
\title{temp title}
\description{}
\usage{
  draw.summary <- function(saved.model, overlay.opts=list(), effect.opts=list(), table.opts=list(), save.dir)
}
\arguments{
  \item{saved.model}{}
  \item{overlay.opts}{}
  \item{effect.opts}{}
  \item{table.opts}{}
  \item{save.dir}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}