\name{easy.mlm.cluster}
\alias{easy.mlm.cluster}
\title{temp title}
\description{}
\usage{
  easy.mlm.cluster(FUNC,FORM, df, model.name = "mdl", outliers.subject = c("fmri", "subject.run", 3), outliers.model = 3, scale.vars = c("fmri", FALSE, 0.9), decompose.vars = c("fmri", "grand", "subject", "subject.run"), save.dir, prefix, cluster.name, cluster.vol, ...)
}
\arguments{
  \item{FUNC}{}
  \item{FORM}{}
  \item{df}{}
  \item{model.name}{}
  \item{outliers.subject}{}
  \item{outliers.model}{}
  \item{scale.vars}{}
  \item{decompose.vars}{}
  \item{save.dir}{}
  \item{prefix}{}
  \item{cluster.name}{}
  \item{cluster.vol}{}
  \item{...}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}