\name{easy.mlm.voxel}
\alias{easy.mlm.voxel}
\title{temp title}
\description{}
\usage{
easy.mlm.voxel( FUNC, FORM, df, model.name = "mdl", outliers.subject = c("fmri", "subject.run", 3), outliers.model = 3, scale.vars = c("fmri", FALSE, 0.9), decompose.vars = c("fmri", "grand", "subject", "subject.run"),  output = c("coef", "aov", "ressd", "outrm", "lsm", "dlsm"), coords, img.dims, save.dir, prefix, ...)
}
\arguments{
  \item{FUNC}{}
  \item{FORM}{}
  \item{df}{}
  \item{model.name}{}
  \item{outliers.subject}{}
  \item{outliers.model}{}
  \item{scale.vars}{}
  \item{decompose.vars}{}
  \item{output}{}
  \item{coords}{}
  \item{img.dims}{}
  \item{save.dir}{}
  \item{prefix}{}
  \item{...}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}