\name{examples}
\alias{examples}
\docType{data}
\title{
Examples of Misclassifications of Units
}
\description{
Some examples of misclassifications of units in clusters of the first and second partitions. The data are the same as used in Cugmas and Ferligoj (2018) (see Figure 4 in the paper) (the exception are examples d, h, l and p which are random and therefore they can slightly differ from those in the paper).
}
\usage{data("examples")}
\format{
The data are in a list format. Each element of the list is a contingency table. The newcomers and outgoers are not denoted.
}
\details{
Element of the list (example in Figure 4 and in Table 3 in Cugmas and Ferligoj (2018))

1 (A), 2 (B), 3(C), 4 (D)

13 (E), 14 (F), 15 (G), 16 (H)

5 (I), 6 (J), 7 (K), 8 (L)

9 (M), 10 (N), 11 (0), 12 (P)
}
\source{
Cugmas, M., & Ferligoj, A. (2018). Comparing two partitions of non-equal sets of units. Advances in Methodology and Statistics, 15(1), 1-21.
}
\keyword{datasets}
