\name{lme.outliers.by.std.resid}
\alias{lme.outliers.by.std.rsid}
\title{Remove outliers from LMER model by standardized residual}
\description{}
\usage{
lme.outliers.by.std.resid(model, data, trim = 2.5)
}
\arguments{
  \item{data}{}
  \item{response}{}
  \item{subject}{}
  \item{trim}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}