\name{lme.outliers.by.subject}
\alias{lme.outliers.by.subject}
\title{Remove outliers from LMER model by subject / scan}
\description{}
\usage{
lme.outliers.by.subject(data, response, subject, trim = 2.5)
}
\arguments{
  \item{data}{}
  \item{response}{}
  \item{subject}{}
  \item{trim}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}