% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{longest_substring}
\alias{longest_substring}
\title{Find Longest Common Substring}
\usage{
longest_substring(a, b, default = NA_character_)
}
\arguments{
\item{a}{a character.}

\item{b}{a character.}
}
\value{
Returns the longest common substring of two character. If multiple longest substrings 
of equal length are found, function will return the leftmost string.
}
\description{
Find the longest common substring in two character strings. For character vectors, 
use \link[=longest_substring_vec]{longest_substring_vec}.
}
