% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{longest_substring_vec}
\alias{longest_substring_vec}
\title{Longest_substring_vec}
\usage{
longest_substring_vec(a, b = NULL, default = NA_character_,
  matrix_out = FALSE, USE.NAMES = FALSE)
}
\description{
A blabla to blablabla
}
