\name{mlm.mediation}
\alias{mlm.mediation}
\title{temp.title}
\description{}
\usage{
  mlm.mediation(FUNC.xm, FIXED.xm, FUNC.xmy, FIXED.xmy, RANDOM, DATA, x.name, m.name, confidence = 95, sensitivity = 0, plot = FALSE)
}
\arguments{
  \item{FUNC.xm}{}
  \item{FIXED.xm}{}
  \item{FUNC.xmy}{}
  \item{FIXED.xmy}{}
  \item{RANDOM}{}
  \item{DATA}{}
  \item{x.name}{}
  \item{m.name}{}
  \item{confidence}{}
  \item{sensitivity}{}
  \item{plot}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}