\name{parse.4d}
\alias{parse.4d}
\title{
Parse 4D Input
}
\description{
Parse 4D Input
}
\usage{
parse.4d(data.4d)
}
\arguments{
  \item{data.4d}{
Full directory listing or list of filenames, or a list of directories and filenames.
}
}
\value{
A list of full file paths to 4D files
}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\note{
mRi internal function
}