\name{parse.model}
\alias{parse.model}
\title{temp.title}
\description{}
\usage{
  parse.model(FUNC, FORM, out.rm.subject, out.rm.trim=matrix(c(3,3), ncol=2), which.effects="all")
}
\arguments{
  \item{FUNC}{}
  \item{FORM}{}
  \item{out.rm.subject}{}
  \item{out.rm.trim}{}
  \item{which.effects}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}