\name{parse.onsets}
\alias{parse.onsets}
\title{
Parse Onset Input
}
\description{
Parse Onset Input
}
\usage{
parse.onsets(data.onsets)
}
\arguments{
  \item{data.onsets}{
Full directory listing or list of filenames, or a list of directories and filenames.
}
}
\value{
A dataframe for generating convolved regressor csv files.
}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\note{
mRi internal function
}