\name{parse.paradigm}
\alias{parse.paradigm}
\title{
Parse Paradigm Input
}
\description{
Parse Paradigm Input
}
\usage{
parse.paradigm(data.paradigm)
}
\arguments{
  \item{data.paradigm}{
Full directory listing or dataframe or matrix.
}
}
\value{
a data dataframe containing all paradigm information.
}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\note{
mRi internal function
}