\name{roi.svd}
\alias{roi.svd}
\title{First, Right Vector from Singular Value Decomposition of ROI Time-Series}
\description{}
\usage{
roi.svd(data.4d, data.mask, num.factors, scan.start.vol=1, save.dir, fname)
}
\arguments{
  \item{data.4d}{}
  \item{data.mask}{}
  \item{num.factors}{}
  \item{scan.start.vol}{}
  \item{save.dir}{}
  \item{fname}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}