\name{run.cluster}
\alias{run.cluster}
\title{temp.title}
\description{}
\usage{
  run.cluster(data.4d, data.paradigm, data.cluster, which.cluster, my.command, my.function, save.dir, file.prefix, num.cores, save.mean.ts=FALSE, verbose=FALSE, ...)
}
\arguments{
  \item{data.4d}{}
  \item{data.paradigm}{}
  \item{data.cluster}{}
  \item{which.cluster}{}
  \item{my.command}{}
  \item{my.function}{}
  \item{save.dir}{}
  \item{file.prefix}{}
  \item{num.cores}{}
  \item{save.mean.ts}{}
  \item{verbose}{}
  \item{...}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}