\name{run.voxel.doPar}
\alias{run.voxel.doPar}
\title{temp.title}
\description{}
\usage{
  run.voxel.doPar(data.4d, data.paradigm, data.mask, my.function, my.command, save.dir, file.prefix, num.cores, log.name, verbose, ...)
}
\arguments{
  \item{data.4d}{}
  \item{data.paradigm}{}
  \item{data.mask}{}
  \item{my.function}{}
  \item{my.command}{}
  \item{save.dir}{}
  \item{file.prefix}{}
  \item{num.cores}{}
  \item{log.name}{}
  \item{verbose}{}
  \item{...}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}