\name{setup.design}
\alias{setup.design}
\title{temp.title}
\description{}
\usage{
  setup.design(trials, n.tr, tr, method = "lsr", nuisance = NULL, noise = NULL, num.cores = NULL, save.design = FALSE, save.name = NULL)
}
\arguments{
  \item{trials}{}
  \item{n.tr}{}
  \item{tr}{}
  \item{method}{}
  \item{nuisance}{}
  \item{noise}{}
  \item{num.cores}{}
  \item{save.design}{}
  \item{save.name}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}