% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{longest_substring_vec}
\alias{longest_substring_vec}
\title{Longest_substring_vec}
\usage{
longest_substring_vec(a, b = NULL, default = NA_character_,
  matrix_out = is.null(b), USE.NAMES = TRUE)
}
\arguments{
\item{a}{a character vector}

\item{b}{a character vector, or \code{NULL} (default) indicating taking \code{a} as \code{b}}

\item{default}{value when no matching substring is found. By default \code{NA_character_}}

\item{matrix_out}{logical. If \code{TRUE}, returns a matrix of common substrings. If \code{FALSE} returns a vector of common substrings.}

\item{USE.NAMES}{logical. If \code{TRUE}, displays names in output. If \code{matrix_out = FALSE}, names of \code{a} will be used.}
}
\description{
A wrapper that vectorizes function longest_substring. Allows an input vector and either a vector or matrix output.
}
